Added comments to many (most) header files.
Added QComboBoxes for TimeZone and DSTRule to LocationDialog.  In addition, the data for the currently-selected City is displayed in  the "Choose coordinates" section, so the user can easily save a modified version of a location.

Reorganized the Main Toolbar layout.

replaced call to updateTime() in buildGUI(), which is needed after all (otherwise, if the program starts up tracking a planet, the initial position is not correct).

svn path=/trunk/kdeedu/kstars/; revision=151507