Added Jupiter's Galilean Satellites (Io, Europa, Ganymede, Callisto).  They are currently drawn as simple circles (and you must be above zoomlevel 5 to see them).  Also, Planet Trails are now drawn in Horizontal-coordinates mode, and the trail color fades gradually into the sky color.
TODO:  There is some problem with the satellite positions in Horizontal coordinates; the "orbital plane" of the moons appears to swing back and forth as Jupiter rises/sets...something to do with EquatorialToHorizontal().  We have the data to support whether the moon is in front of/behind Jupiter...should take this into account when drawing.  Option to label the moons.

svn path=/branches/make_kdeedu_cool/kdeedu/kstars/; revision=184915