Added KSPopupMenu class, derived from KPopupMenu.  All pop-menu code is now kept in this class instead of SkyMap.
Fixed planet position angles.  The PA was defined relative to
screen X,Y coords instead of RA,Dec.  This required a cascade of
fixes where we depended on the wrong definition of PA.

Added EquatorialToEcliptic() call to Moon's findPosition()

Infoboxes are now "sticky" on startup if they are up against the
window edge.

svn path=/trunk/kdeedu/kstars/; revision=210414