Altitude plotter improvements.  Added a "Clear Fields" button. setTrapReturnKey(true) for each lineedit, so window doesn't close prematurely.  Highlighting a name in the listbox now highlights the corresponding curve in the plot.
When Return is pressed, it will first attempt to recognize the text
in the nameBox as an existing object name, and add the object if it is
recognized.  Then, if the name is not recognized and the raBox and
decBox have valid text, it will add a custom object to the plot.
(previously, this order was reversed; it would add a custom object
if all three fields were filled.  This led to the possibility of
easily adding bogus duplicate objects to the plot).

svn path=/trunk/kdeedu/kstars/; revision=213763