Working toward a better separation of the "backend" from the GUI frontend. The motivation for this is to add a command-line mode to generate a skymap image based on the current settings stored in the kstarsrc file.  This will let people use KStars to dynamically generate desktop wallpapers, for example.
To this end, I have removed usage of a pointer to the KStars object
wherever possible.  KStars is the class for our main window, and also has
been the container for core parts of the program, such as KStarsData,
KStarsOptions, SimClock, etc.  To separate the backend, I have made
KStarsData the keeper of these core non-GUI objects, which will allow us
to avoid direct usage of the KStars object in functions that have nothing
to do with the GUI.

All solar-system classes passed a pointer to KStars in their constructor;
these have been changed to KStarsData pointers.  (In doing this I noticed
that most of these classes didn't even use this pointer, so I am
investigating whether we can remove the KStarsData pointer altogether).

The functions in kstars_options_handling (loadOptions() and saveOptions())
are now members of KStarsData.  The existing KStarsData::saveOptions()
function has been renamed restoreOptions(), which makes more sense since
its function is to restore the original options if "Cancel" is selected in
the ViewOsDialog).  One complication: saveOptions() now takes an optional
KStars pointer, because some of the options are taken directly from GUI
properties, such as the window size and the Focus point.

Note that I also added the necessary code to main.cpp for the
"--dump" command-line option, but it currently does not do anything :)
CCMAIL: kstars-devel@lists.sourceforge.net

svn path=/trunk/kdeedu/kstars/; revision=228370