More work to separate backend from GUI.  Moved the SimClock object from KStars to KStarsData.  We retain a convenience pointer in KStars (but it has been renamed from getClock() to clock() ).
Also, moved intial call to setFullTimeUpdate() from
KStarsData::slotInitialize() to KStars::dataInitFinished().  This is a
trivial little function, didn't make sense that it had its own
entry in the splashscreen (and the splashscreen message was "Starting
Clock", which is a very bad description for setFullTimeUpdate() ).

Removed the KStars pointer from SimClock.  It was needed by the DCOP
function SimClock::setClockScale() to call SkyMap::slotClockSlewing()
directly; instead we now simply connect the SimClock::scaleChanged()
SIGNAL to this SLOT.
CCMAIL: kstars-devel@lists.sourceforge.net

svn path=/trunk/kdeedu/kstars/; revision=228919