AltVsTime tool improvements: + Restored sidereal time labels on top axis + The time axis is now centered on midnight local time, rather than noon + Restored the green background color indicating points below horizon + Added blue background color indicating daylight hours (including twilight gradients) + Added a vertical dotted line that indicates the current time. + Added a POLYGON PlotObject type (not currently used)
TODO:
If we want to keep the vertical line indicating the current time,
we should perhaps add a time edit widget to the Date/Location tab.

Note that in order to do the fancy background colors, elts now uses a
subclass of PlotWidget (called AVTPlotWidget).  The only difference is the
paintEvent().  The class definition is in elts.h/elts.cpp.

CCMAIL: kstars-devel@lists.sourceforge.net

svn path=/trunk/kdeedu/kstars/; revision=234508