First use of the libkdeedu/kdeeduplot library in KStars.  From now on, KStars will depend on having this (very small) library installed.  All plot-like tools (Alt vs. Time, Jupiter Moons, and Solar Sytem Viewer) now use KStarsPlotWidget, derived from KPlotWidget.
The Jupiter Moons tool now has some keyboard interactions: Press +/- to
change the stretch on the time axis.  Press "[" / "]" to adjust the
zeropoint of the time axis.  (I wanted to use Arrow keys instead of
brackets, but it turns out this is a bit complicated, because arrow key
events get eaten by the dialog...there is a way around it with
EventFilters, but I haven't dealt with it yet).

TODO:

the time axis of the Jupiter Moons tool currently displays offset
from now, in days.  This is not very useful; I want to change it to the
actual dates.

Add name labels to planets in the Solar sysytem viewer.

CCMAIL: kstars-devel@lists.sourceforge.net

svn path=/trunk/kdeedu/kstars/; revision=237010