Changed the way the zoom level is handled in the code.  Previously, we had a hard-coded array of doubles that set the pixel scale (called pixelScale[]), and zooming in and out would [dec/inc]rement the index of the array.  We now simply store a double for the current pixelScale (called ZoomFactor), and zooming in/out changes this value by multiplying/dividing by a preset factor DZOOM.
At this point, DZOOM is #defined in kstarsdata.h to be 1.1; it could
conceivably be an Advanced view option.  The smaller DZOOM is, the smoother
zooming will be (but there will also be more steps).

Doing Zoom this way will allow us to set the ZoomFactor arbitrarily, so that
a "Zoom-to-box" function can be implemented.  It would also be nice to have
a DCOP function "setZoom( double )".

svn path=/trunk/kdeedu/kstars/; revision=239275