Added "Export Sky Image..." item to File menu, which lets you save the current sky map as an image file.  It is capable of saving the file to a remote location.
The "Run Script..." action can now access remote kstars scripts. There
is a scary-worded warning if you try to run a remote script, with options
to cancel the download, or save it locally without running it.  (I also
note that all scripts are validated before they are run; if there's
anything in a script besides kstars DCOP commands, it will not be
executed...this isn't new to this commit, but it bears pointing out).

Added PNM and BMP to the list of image formats to which command-line image
dump mode can export.

CCMAIL: kstars-devel@lists.sourceforge.net

svn path=/trunk/kdeedu/kstars/; revision=242837