GUI modified to add the batch Mode which allows to compute ephemeris from a collection of input parametres in an input file and write the results to an output file.
The GUI uses now a container made of two tabs, as in other modules which
have already been modified.

The original widget is contained in the Interactive Mode tab, while the
Batch Mode tab contains the wigdets to make copmutations using parametres
in a file. The code for the batch mode is not implemented yet, so only
the Interactive Mode is operative.

CCMAIL: kstars-devel@lists.sourceforge.net

svn path=/trunk/kdeedu/kstars/; revision=242901