Well, it took all day, but I fixed the Moon bug.
It all started when Jasem noticed that the Moon "jumped" when its popup
menu was shown.  I tracked the problem to the calculation of rise/set
times, which reset the moon's position to geocentric coordinates (i.e.,
without correcting for the fact that the location is on Earth's surface,
6500 km from the center).

In fixing the problem, I decided to finally apply this "figure of the
Earth" correction to all solar system bodies, not just the Mooon (even
though the effect is close to negligible for other bodies, since they
are much further away).  This caused the usual cascade of changes in
many parts of the code:

Prior to these changes, each subclass of KSPlanetBase (KSPlanet,
KSPluto, KSSun, KSMoon, KSAsteroid, and KSComet) had overloaded its own
unique 'findPosition()' function.  This function has been renamed
'findGeocentricPosition()', and is now a protected function (it
shouldn't be called directly from outside the member functions anymore;
instead call the new findPosition()...see below).

Moved KSMoon's functions 'findPosition(KSNumbers *num, dms *lat, dms
*LST, KSPlanetBase *Earth)' and 'localizeCoords(dms *lat, dms *LST)' to
KSPlanetBase.  These are *not* overloaded by the subclasses.
findPosition calls the overloaded findGeocentricPosition(), and then
calls localizeCoords() to correct for figure-of-the-Earth.  Since
localizeCoords() is only used in findPosition(), it is private.  These
moves allow the figure-of-the-Earth correction to be applied to all
solar system bodies.

moved KStarsData::isSolarSystem(SkyObject*) to
SkyObject::isSolarSystem().

several of the functions related to rise/set times in SkyObject had
tweaks to their arguments.  Basically, instead of passing latitude and
longitude as dms pointers, we now simply pass a GeoLocation pointer.

Added rearth() and setRearth() to KSPlanetBase. These read and set the
distance of the body from Earth (which we need for localizeCoords()).
setRearth( KSPlanetBase *earth ) will compute the distance;
setRearth(double) will simply set it (in AU).  setRearth(*earth) doesn't
do anything for the Moon, because its Rearth is computed as part of its
findGeocentricPosition()

Removed now-redundant KSMoon::distance()

Added '#define AU_KM 1.49605e8' to kstarsdata.h, the number of km in one
AU.

PlanetCatalog::findPosition( KSNumbers *num ) moved to findPosition(
KSNumbers *num, dms *lat, dms *LST )

CCMAIL: kstars-devel@lists.sourceforge.net

svn path=/trunk/kdeedu/kstars/; revision=243219