The star catalog is now derived from the Hipparcos/Tycho Star catalog, rather than the SAO star catalog.  The primary reason for this change is that the Hipparcos/Tycho catalogs have been confirmed to be in the public domain, while the status of the SAO catalog was stubbornly ambiguous.  In addition, the change provides additional benefits:
+ sub-arcsecond precision in positions of stars
+ Inclusion of proper motion and parallax distances for many stars
+ Indication of multiplicity and variability of all stars

The StarObject class now stores the proper motion, parallax,
multiplicity, and variability data, but we don't yet make use of it
in the program.

In preparing the catalog for KStars, I had to do a lot of sorting and
reformating, and I also had to assign the star names by hand.
Since this was a lot of work, I put a copyright notice at the top of
hip001.dat, indicating that the files are protected by the OPL.

In addition, I added a feature to the data/Makefile.am that will
detect and remove any deprecated data files.

CCMAIL: kstars-devel@lists.sourceforge.net, jfs@computer.org, bab@debian.org

svn path=/trunk/kdeedu/kstars/; revision=243830