Implemented Pablo's suggestion of crosshairs in the AltVsTime plot.  There is a vertical line and a horizontal line in the plot which intersect at the current cursor position.  It works well, except for some reason I don't understand, the lines are only updated if you hold down the mouse button while moving the mouse.  I used mouseMoveEvent() for this, and I can't see why the event is (apparently) not firing when the mouse button is not pressed.  Please take a look at altvstime.cpp and see if you can figure out what is happening...
Actually, maybe we don't want to fix this.  It might be nice to be able to
leave the cursor lines wherever you want them; the current setup is a
natural way to do that.  You can also click anywhere in the plot widget,
and the lines will be drawn at the cursor position (I just added a
mousePressEvent(), which calls mouseMoveEvent(e) ).

CCMAIL: kstars-devel@30doradus.org

svn path=/trunk/kdeedu/kstars/; revision=244134