Implement wish #63365.  Cardinal compass point labels (N, NE, E, SE, etc) are now drawn along the horizon line, regardless of whether the opaque ground is drawn or not.  When the opaque ground is not drawn, the compass labels use the horizon color rather than the compass color, because the compass color is often very similar to the sky color, which would make the labels invisible if that color were used.
Note that compass labels are still only drawn when using Horizontal
coordinates.  The problem is that the compass labels need to be rotated
when used with Equatorial coordinates.  I know how to do that, but I am
going to wait until I add labels to the coordinate grid and ecliptic, so
that they can all use the same code.

CCMAIL: 63365-done@bugs.kde.org, kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=246460