I used valgrind with memcheck to check for leaks, unintilized values, and invalid read/writes
Valgrind complained about some leaks and uninitlized values which I fixed. One of which resided in libkdeedu/kdeeduplot/kplotwidget. The QPixmap buffer was not freeed and so I add if (buffer) delete buffer in the destructer.

This was not an extensive test, but we should certainly use valgrind again for kstars-wide checkup

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=247128