Increased the number of stars with genitive names from 220 to 1500.
Genetive names are constructed by using a Greek letter to indicate
the star's brightness rank within its constellation, and combining this
with the possessive form of the constellation name, in Latin.  For
example, the brightest star in Lyra is named Vega, and Vega's genitive
name is therefore "alpha Lyrae".

I also modified the way genitive names are used in the program.
Previously, the full genitive name was recorded in the star database,
and stored in the StarObjects.  Now, an abbreviated form is used in the
database and StarObjects (i.e., "omi Ori"), which is interpreted
into its full form ("omicron Orionis") by the StarObject::gname() function.
gname() calls both StarObject::greekLetter() and StarObject::constell() to
interpret the two parts of the genitive name abbreviation.

TODO:
Might be nice to have an option that draws greek letters on the map
instead of (or in addition to) star name labels.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=247743