The atmospheric refraction at the horizon shifts altitude by - 34 arcmin = 0.5667 degrees. This value changes if the observer is above the horizon, or if the weather conditions change.
For the sun we have to add half the angular sie of the body, since
the sunset is the time the upper limb of the sun disappears below
the horizon, and dawn, when the upper part of the limb appears
over the horizon. The angular size of the sun = angular size of the
moon = 31' 59''.

So for the sun the correction is = -34 - 16 = 50 arcmin = -0.8333

This same correction should be applied to the Moon however parallax
is important here. Meeus states that the correction should be
0.7275 P - 34 arcmin, where P is the moon's horizontal parallax.
He proposes a mean value of +0.125 degrees if no great accuracy
is needed.

 Changed "if" condition, because I had exchanged the Sun and the stellar
cases and added a bigger comment explaining why we use these values.

CCMAIL:kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=247995