Changed function GSTat0hUT. This function is called by GSTtoUT and the latter by LSTtoUT and UTtoLST. Therefore the code here is used everywhere a time transformation is needed.
I have added a correction so that the function returns the apparent GST instead
of the mean GST. The correction is made by adding the term
dEcLong*cos(obliquity) where dEcLong = the change in Ecliptic Longitude
due to nutation, and obliquity is the obliquity of the ecliptic. The correction
may amount 0.3 seconds of time.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=248547