Adding Customizable "Field of View" symbols (f/k/a Target Symbols).  Note that this appears on the 3.2 Planned Features document.
A Field of View symbol is defined by a very simple class named FOV. FOV stores the symbol's size in arcmin, its color, and its shape (0=square, 1=circle, 2=crosshairs, 3=bullseye).  It has accessor methods for these variables, and a draw() function.

There are three default FOV symbols in the Settings->FOV Symbols menu ("7x35 Binoculars", "One degree", and "HST/WFPC2") and an "Edit FOVs..." option, which opens a dialog to edit, remove or add new FOV's.

The FOV's are defined in fov.dat, which is stored locally, at $KDEHOME/share/apps/kstars/fov.dat.  If this file doesn't exist, a default version is created.  This way, the user can remove the defaults, if they wish.  After 3.2, I'd like to do this with color schemes as well.

Creating a new FOV is easy.  You first define the angular size of the symbol, in arcmin.  You can either enter a value manually, or fill-in telescope/eyepiece or telescope/camera parameters to compute a value.  Next, you set the shape and color for the symbol, and name it.  That's it.

The FOV parameters (size, shape, color) are stored separately in the config file.  These can be manipulated through the DCOP function "changeViewOption()", and they are already incorporated into the ScriptBuilder tool.

TODO:
+ The FOV symbol currently in use should be pre-selected in the
  list when the FOV dialog is opened.
+ If the FOV Dialog is closed with a FOV highlighted in the list,
  activate that FOV in the map.
+ The FOV "preview" in the New FOV dialog does not always refresh
  to show the current parameters.

CCMAIL: 63371-done@bugs.kde.org, kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=249694