Fixed a crash condition when clicking "edit" in the FOV dialog when no items are selected. I also noticed that in the new FOV dialog, the word "length" appreaed as "lengt" only, I thought first it was a type but when I opened the ui file, it wasn't a type but rather a layout issue. Since different users have different font and resolution settings, it's better to encapsulate all widgets in layouts.
The new ui can grow/shrink in size without problem in Qt designer. However, because FOVDialog is a child of KDialogBase, the shrinking of the FOV ui is not as pretty. This shrinking problem exists with the old ui as well.

The FOV feature is great Jason! When I created a new FOV and then switched to it from the menu, I sort of expected KStars to zoom to the new FOV for the current focus in the skymap (or center). The zoom should be a little larger than the FOV symbol itself so the user can see the symbol. If a user switches to very small FOV (like that of HST), they don't know the change because the symbol is tiny, so maybe the autozoom when switching FOVs would resolve that issue as well.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=249788