Added a Zoom-dependent faint magnitude limit for Deep-sky objects, similar to what we already have for stars.  Note that this is in the 3.2 Planned Features list.
With this change, fainter Deep-sky objects are not visible unless you zoom in.  This makes the zoomed-out sky look more like what you can see with your eyes, and the zoomed-in sky more like a telescopic view.  If you prefer the old behavior (all deep-sky objects drawn all the time), just set both parameters to the maximum setting (16.0).

As with stars, the zoom-dependent limit is controlled by a pair of magnitudes: the draw limit when zoomed in and when zoomed out.  Widgets to adjust these parameters have been added to the ViewOpsDialog.  These new options have also been added to the ScriptBuilder.

CCMAIL: kstars-devel@kde.org, 61682-done@bugs.kde.org

svn path=/trunk/kdeedu/kstars/; revision=250133