Bug #64015 reported that non-integer time zone (TZ) offsets did not function at all in KStars (for example, some locations in AU have TZ=9.5, but the Local Time in KStars was always UT+9:00).  However, looking at my local working copy, the TZ offset appears to be working correctly, so it must have been fixed at some point :)
However, there was still a related usability problem in the "Set Location"
dialog: the TimeZone combobox showed only integer TZ values between -12
and 12.  So if you selected a location with, say, TZ=9.5 it would be
displayed in the combobox as "9".  For existing locations, this is just
a cosmetic issue, but it also makes it impossible to add a new location
with a non-integer TZ offset.

The combobox is now editable (so you can enter any value you want, or
select from the list).  In addition, the TZ box is initially populated
with all unique TZ values present in the Cities.dat database (including
any already-defined custom locations).

CCMAIL: 64015-done@bugs.kde.org, kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=250621