Bugfix: when tracking on empty sky, clicking on an object would cause the focus to recenter on the clicked object at the next skymap update.
This is because SkyMap::updateFocus() used clickedPoint() to determine
the coordinates of the focus when centered on empty sky.

To fix, I added a SkyPoint member named FocusPoint to store the focus
position when tracking on empty sky.  Accessors are focusPoint() and
setFocusPoint().

Skymap now has the following SkyPoint members: MousePoint, ClickedPoint,
Focus, Destination, and FocusPoint (and SkyObject pointers ClickedObject
and FocusObject).  This is a bit complicated, but I think they are all
necessary.  Perhaps FocusPoint and Destination can be combined, but
this should wait until after 3.2.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=251155