Fixed handling of Image/Info URL files.  The files image_url.dat and info_url.dat contain data for web links attached to the popup menus of various objects.
The first time KStars is run, the globally-installed
files are copied into the user's KDEHOME directory,
so they are free to add/modify/remove links as they
see fit.  However, this means that if the global
version of the file is updated, the user will not
see the updates, because KStars only looks at the
global file if the local file does not exist.

This is now fixed.  KStarsData::openURLFile()
compares the timestamps of the global and local
versions.  If the global version is newer, then
it is checked for new or duplicate lines.  When
checking lines, only the object-name and text-label
fields are considered, not the URL itself.  This
is done so that if a link's URL is updated, it is
still considered a duplicate of the original line.
New lines in the global file are appended to the
local file.  When duplicate lines are found, the
line from the global file replaces the line in the
local file.

Please test this on your system using the recently
updated image_url.dat.  Before running
kstars, your local image_url.dat should have ~440
lines. Afterwards, it should have ~650 lines.

CCMAIL: kstars-devel@kde.org
----------------------------------------------------------------------

svn path=/trunk/kdeedu/kstars/; revision=254958