Fixed crash condition: requesting an image dump with width less than 380 caused a crash.  Traced the problem to the definition of the xscale/yscale parameters in SkyMap::exportSkyImage():
  double xscale = pdm.width() / width();
  double yscale = pdm.height() / height();

The width() and height() functions return integers, so if pdm.width() <
width() then xscale is incorrectly set to 0.0.  This eventually
propogates to a div-by-zero.  Fixed by wrapping the width/height in
"double()".

The reason this only manifests when width is smaller than 380 is that the
skymap widget gets resized to 380 when it is small, but the
PaintDeviceMetric (pdm) remains at the original value.  For widths above
380, width()==pdm.width() and xscale==1.0.

Thanks to Tobias Koenig for pointing out the problem.

CCMAIL: kstars-devel@kde.org
CCMAIL: tokoe@kde.org

svn path=/trunk/kdeedu/kstars/; revision=255741