Dynamic loading of stars works now.  There were two fixes required:
(1) Loading more stars by increasing the faint limit.  This code was in place already, but in
FileSource::sendTo() it parses the star's magnitude from the data file, and this was not updated
when the star catalog was switched to the Hipparcos catalog, so it was looking at the wrong
position in the line.

(2) Removing stars when decreasing the faint limit.  Previously, stars were not unloaded from the
program, just the faint limit was changed.  Now unused faint stars are actually removed, thus
giving users the ability to free up memory by decreasing the faint limit.  Also, if any removed
star is named, its entry is removed from the list of named objects.  I had to add a function
ObjectNameList::remove(const QString &name) for this.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=256872