Fixing some issues as I was going through Jason's QA list:
1. Several i18n fixes.
2. Adding new INDI hosts triggered both connect and disconnect buttons, it's fixed now (only connect is enabled for new added hosts).
3. Adding/Editing/Removing links was broken. We make all modifications against the local image_url.dat and info_url.dat now (no more myinfo_url.dat myimage_url.dat).

New entries are simply appended to the list. edited entries are deleted and then appended to the list. deleted entries are simply removed from the list. The only issue I can see here is that if a user upgrades and gets new image_url.dat info_url.dat then all the 'default' links he deleted will be back again. This is good or bad depending on how you look at it. On one hand, this is one way to allow the user to retrieve deleted links. On the other hand, the user might not want some of the default links to come back after an upgrade. What do you think?

4. INDI version update: updated the version number for indi devices where appropiate based on user feedback.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=257153