Two fixes related to rise/set/transit times:
(1) The rise/set/transit times are displayed in three places: the
popup menu, the Details dialog, and in the WUT tool.  In the WUT,
if the transit or set time is before the rise time, then the
time is recomputed for the following day.  I added this
provision to the popup menu and the Details dialog as well, so
that the set and transit times are now consistent.

(2) Determining the rise/set/transit times is an iterative
process that recomputes the position of the object for the Julian
Day corresponding to the first-guess rise/set/transit time, and
then determines the rise/set/transit time for that revised
position.  The function
SkyObject::getNewCoords(jd,jd0,geo) is supposed to then restore
the object's position to the current JD, under the assumption
that the first parameter ("jd") is the current Julian Day value.
However, this assumption was violated in the WUT, resulting in the
restored position being incorrect.  This was especially a problem
for the Moon; its restored position was off by over 10 degrees.

I have replaced SkyObject::getNewCoords() with
SkyObject::computeCoordsForJD(jd, geo).  This function simply stores
the object's original position in a local variable and resets it at
the end, rather than recomputing it.

Fix (1) is backported to KDE_3_1_BRANCH and debian_branch; Fix (2)
is not, because there is no WUT in those branches, so the code
(while fragile) is not broken.

TODO: the rise/set/transit times can still be in disagreement by +/- 1
minute between the three locations; this still needs investigating.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=259001