Improvements to splashscreen, to fix two refresh issues:
1. occassionally the background of the text region would become light
grey, making the white text impossible to read.  Fixed by adding an
"inactive background" color to the widget palette.

2. Two of the load items (stars and deep-sky) take a significant amount of
time, so a long time passes during loading with no apparent progress in
the splashscreen.  Added "N%" text to the loadStarData() and
loadDeepSkyData() functions, which update whenever a new data file is
opened (127 and 14 files for stars and deep-sky, respectively).  I timed
the functions with and without this addition, and the additional updates
only add about 1% to the startup time.

This has the additional benefit that the splashscreen is refreshed more
often, so if another window is placed in front of it temporarily, there
will no longer be a long interval during which the splashscreen is
completely blank.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=259452