two bug fixes:
1. When tracking an object with Horizontal coordinates and a large
timestep, the tracking position lagged behind the object.  What we had
done was call 'setDestination( focusObject() )' followed immediately by
'setFocus( destination() )'.  However, setDestination() triggers a signal
connected to the slewFocus() slot, which was apparently taking so long
that it never got to the 'setFocus( destination() )' before it received
the next setDestination().  Anyway, that's probably confusing, but the
short version is, we now instead call 'setFocus( focusObject() )' followed
by 'setDestination( focus() )'.  This works much better, because when
slewFocus() is called, focus and destiantion are guaranteed to be equal,
so it returns immediately, and the tracking stays locked on the object.

2. When tracking on a solar system body and then making a discontinuous
time change (with either the "Set Time" or "Set Time to Now" actions), the
body's trail is now cleared so that you don't get a big, ugly line segment
between the previous and new positions.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=264347