Allows to convert between equatorial and galactic coordinates using any epoch for the former coordinates.
This option was not coded before resulting in a bug when the epoch of the
equatorial coordinates was not B1950. To avoid the bug we used a
workaround: we locked the epoch in the UI file to always be 1950.0.

With the present change we have unlocked the epoch field in the UI file
and one can choose any epoch for the equatorial cooordinates. The change
consists on adding precession from the input epoch to B1950. That is, if
one goes from Equatorial to Galactic. the order is:
 - precession of equatorial coordinates from the given epoch to B1950
 - conversion from equatorial to galactic coordinates
If one goes from Galactic to Equatorial, the order is:
 - conversion from galactic to equatorial coordinates
 - precession of equatorial coordinates from B1950 to the given epoch.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=274712