Fixing bug #71520 ("Strange behavior of What's up tonight").  Problem traced to the fact that planet positions were not recomputed when user changed the Date inside the tool.  In order to fix, I am using the SkyObject::computeCoordsForJD() function.  I also had to make this function public.
Incidental fix:  wutdialog.cpp had code that was meant to always exclude
the Sun from the list of visible objects, but the code only worked for
English, because it compared the translated name to the untranslated name.
In addition, the same code also excluded the Moon.  However, the Moon
should NOT be excluded like this, so I removed that part.

(note that part 3 of this bug report, regarding the date format displayed
in WUT, was found to be invalid).

Thank you for the detailed bug reports!

CCMAIL: kstars-devel@kde.org
CCMAIL: 71520-done@bugs.kde.org

svn path=/trunk/kdeedu/kstars/; revision=275815