Fixing last portion of compound BR #71520.  The reporter found that the WUT tool did not report that the Moon and Venus were up on the night of 26 Dec. 2003.  He provided photographic evidence that they were, in fact, visible that night :)
The problem was in our definition of what to consider "up" on a given
night.  Naively, one might just consider whether the object is above the
horizon at any time between sunset and sunrise.  However, what we really
want to know is, can one *observe* a given object from a given location,
on a given night?  In this case, we want to know if the object is above
the horizon between the end of evening twilight and the beginning of
morning twilight, not the moments of sunset and sunrise.

Twilight is defined to end or begin when the Sun is at a certain negative
altitude below the horizon.  There are three kinds of twilight: civil
twilight (alt=-6 deg), nautical twilight (alt=-12 deg) and astronomical
twilight (alt=-18 deg).  In KStars, we had been using -20 degrees, which
is close to astronomical twilight.

However, this is probably too strict, especially for planets and the moon
which are often easily visible low in the sky during twilight.  We are now
using the definition of civil twilight, -6 degrees, to determine whether
an object is visible.

Wow, four paragraphs to explain a one-line change in the code.  :)  Sorry
abot that!

CCMAIL: 71520@bugs.kde.org
CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=276066