Fixing bug in the AltVsTime tool that Pablo pointed out: the altitude curves of custom objects did not update when changing the date/location.
Also, I noticed the Epoch box was not being used to set the epoch for new
custom objects.  Now, if the box is not blank when adding a custom object,
the coordinates will be precessed from the entered epoch to J2000.  Thus,
it assumes that a blank epoch means J2000 (perhaps we should instead
assume the epoch of the tool's current date?  any opinions?)

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=276830