Fixing Bug #71829: star is not i18n
It's a bit complicated, but essentially:  We keep several name-strings for
each object, including the untranslated English Name [name()], the
translated name [translatedName()], and a long name [longName()].

We display the "long name" of sky objects in the status bar and the
popup menu.  In the case of unnamed stars, the "long name" was being
assigned the non-localized name "star".  We now are using
"translatedName()" in SkyObject::setLongName(), so it
should work.  Please test.

Incidental fixes: Fixed three instances where we compare the untranslated
name to "i18n("star")" instead of just "star" in starobject.cpp and
kspopupmenu.cpp.

CCMAIL: 71829-done@bugs.kde.org
CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=277683