Implemented proper motion of stars.  That was easy, should have done it a long time ago! :)
I added StarObject::updateCoords(), which overrides the base version in
SkyPoint.

You can check it out by centering on a high PM star like Sirius or
Procyon, and either setting the date to something like year 7000 or
setting the timestep to 50 yrs or so.

Note that constellation lines are not tied to the stars; this is going
to change soon :).

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=283733