Added funcionality to measure angles bewteen two points in the sky map. To begin measuring one has to right click on any object of the skymap and select the appropiate item in the popup menu. This is coded in KSPopMenu. The coordinates of the clicked object are stored in PreviousClickedObject and a bool variable is set to true indicating that we are in measuring angle mode. This part is coded in SkyMap.
To finish the measurement one has to left click on any point of the sky map
and the result will be shown in the status bar together with the name of the
clicked object. This part is coded in skymapevents.cpp. The computation of
the angular distance is done here, the result is shown in the status bar and
the masuring angle mode is set to false.

Still have to add a way to indicate that we are in measuring angle mode. I
want to draw a line from the first point to any place of the sky map. That
line would be updated as we move the mouse while the mode is on. As soon
as we click on the second point, the line will disappear and the result
will be shown in the status bar.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=284708