Reworked the angular distance measurement mode. Now the angle measuring mode is activated from the popupmenu or clicking on the skymap and pressing CTRL+[. The angle mode is finished from the popupmenu or clicking on the skymap and pressing CTRL+].
When the angle measuring mode is on, a line is drawn from the first
point to the second one. When the angle mode is finished the result is shown
in the status bar and the line disappears.

I have suppressed from skymapevent the event triggered from the left button
click. To draw the line I defined a function called drawRuler in skymapdraw as
Jason told me. That works nicely since it is very easy to implement it.

TODO:
 - There are some errors when plotting the line. I do not know why yet, but
   the start point is offset by a small quantity from the clicked point.
 - I do not like the labels I have written in the KSPopMenu class. I will try
   to fix that, or better, somebody speaking good english (Jason :-) ) writes
   some better labels
 - Investigate if the line can flash or if another color is nicer.
 - See if it is possible to avoid clicking first with the left button and then
   press CTRL+[. It would be easier to just do CTRL+[ and the position where
   the mouse is hovering on is read.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=285577